move handlers into a separate file
